# doc/api.yaml
openapi: 3.0.0
servers:
  - url: "http://localhost:3000"
info:
  title: WASAText
  description: API specification for the WASAText project (Web and Software Architecture).
  version: 1.0.0

# Definizione dei tag per raggruppare le chiamate nella documentazione
tags:
  - name: Status
    description: System status endpoints
  - name: Auth
    description: Simplified authentication (Login)
  - name: User
    description: User profile management
  - name: Conversations
    description: Chat management (Direct and Groups)
  - name: Messages
    description: Message handling
  - name: Groups
    description: Specific group management

# Sicurezza globale: Bearer Token richiesto ovunque tranne dove sovrascritto (es. login)
security:
  - bearerAuth: []

paths:
  # --- System Status ---
  /status:
    get:
      tags: ["Status"]
      summary: Returns the current status of the server
      operationId: getStatus
      security: [] # Endpoint pubblico
      responses:
        '200':
          description: Server is running correctly
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    description: Server status message
                    example: "alive"
                    pattern: "^[a-z]+$"
                    minLength: 1
                    maxLength: 20

  # --- Authentication (Simplified Login) ---
  /session:
    post:
      tags: ["Auth"]
      summary: Logs in the user (Simplified)
      description: >
        If the username exists, logs in. If not, registers and logs in.
        Returns the user identifier to use as Bearer token.
      operationId: doLogin
      security: [] # Endpoint pubblico
      requestBody:
        description: User details (username only)
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  description: The username to login with
                  example: Maria
                  minLength: 3
                  maxLength: 16
                  pattern: "^[a-zA-Z0-9]+$"
      responses:
        '201':
          description: User log-in action successful
          content:
            application/json:
              schema:
                type: object
                required:
                  - identifier
                properties:
                  identifier:
                    type: string
                    description: The user ID to be used as Bearer token
                    example: "u123456"
                    minLength: 1
                    maxLength: 64
                    pattern: "^[a-zA-Z0-9_-]+$"
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # --- User Profile ---
  /users/me/name:
    put:
      tags: ["User"]
      summary: Update my username
      operationId: setMyUserName
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  description: New username
                  minLength: 3
                  maxLength: 16
                  pattern: "^[a-zA-Z0-9]+$"
      responses:
        '200':
          description: Username updated
          content:
            application/json:
              schema:
                type: object
                description: Empty success response
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /users/me/photo:
    put:
      tags: ["User"]
      summary: Update my profile photo
      operationId: setMyPhoto
      requestBody:
        required: true
        content:
          image/png:
            schema:
              type: string
              format: binary
              description: The photo image data
              minLength: 1
              maxLength: 10485760 # Max 10MB
              pattern: "^.*$"
      responses:
        '200':
          description: Photo updated
          content:
            application/json:
              schema:
                type: object
                description: Empty success response
        '500':
          $ref: '#/components/responses/InternalServerError'

  # --- Conversations ---
  /conversations:
    get:
      tags: ["Conversations"]
      summary: Get list of conversations
      operationId: getMyConversations
      responses:
        '200':
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConversationSummary'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /conversations/{conversationId}:
    parameters:
      - $ref: '#/components/parameters/ConversationId'
    get:
      tags: ["Conversations"]
      summary: Get specific conversation details and messages
      operationId: getConversation
      responses:
        '200':
          description: Conversation details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationDetail'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # --- Messaging ---
  /conversations/{conversationId}/messages:
    parameters:
      - $ref: '#/components/parameters/ConversationId'
    post:
      tags: ["Messages"]
      summary: Send a message
      operationId: sendMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - content
              properties:
                content:
                  type: string
                  description: Message text
                  minLength: 1
                  maxLength: 2000
                  pattern: "^.*$"
      responses:
        '201':
          description: Message sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /conversations/{conversationId}/messages/{messageId}:
    parameters:
      - $ref: '#/components/parameters/ConversationId'
      - $ref: '#/components/parameters/MessageId'
    delete:
      tags: ["Messages"]
      summary: Delete a message
      operationId: deleteMessage
      responses:
        '204':
          description: Message deleted
        '403':
          description: Forbidden (not your message)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /conversations/{conversationId}/messages/forward:
    parameters:
      - $ref: '#/components/parameters/ConversationId'
    post:
      tags: ["Messages"]
      summary: Forward a message
      operationId: forwardMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - messageId
              properties:
                messageId:
                  type: string
                  description: ID of message to forward
                  minLength: 1
                  maxLength: 64
                  pattern: "^[a-zA-Z0-9_-]+$"
      responses:
        '201':
          description: Message forwarded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # --- Reactions ---
  /conversations/{conversationId}/messages/{messageId}/comment:
    parameters:
      - $ref: '#/components/parameters/ConversationId'
      - $ref: '#/components/parameters/MessageId'
    put:
      tags: ["Messages"]
      summary: React to a message
      operationId: commentMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - emoji
              properties:
                emoji:
                  type: string
                  description: The emoji character
                  minLength: 1
                  maxLength: 4
                  pattern: "^.*$"
      responses:
        '200':
          description: Reaction added
          content:
            application/json:
              schema:
                type: object
                description: Success
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: ["Messages"]
      summary: Remove reaction
      operationId: uncommentMessage
      responses:
        '204':
          description: Reaction removed
        '500':
          $ref: '#/components/responses/InternalServerError'

  # --- Groups ---
  /groups/{groupId}/members:
    parameters:
      - $ref: '#/components/parameters/GroupId'
    put:
      tags: ["Groups"]
      summary: Add user to group
      operationId: addToGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - userId
              properties:
                userId:
                  type: string
                  description: User ID to add
                  minLength: 1
                  maxLength: 64
                  pattern: "^[a-zA-Z0-9_-]+$"
      responses:
        '200':
          description: User added
          content:
            application/json:
              schema:
                type: object
                description: Success
        '403':
          description: Not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /groups/{groupId}/members/me:
    parameters:
      - $ref: '#/components/parameters/GroupId'
    delete:
      tags: ["Groups"]
      summary: Leave a group
      operationId: leaveGroup
      responses:
        '204':
          description: Left group
        '500':
          $ref: '#/components/responses/InternalServerError'

  /groups/{groupId}/name:
    parameters:
      - $ref: '#/components/parameters/GroupId'
    put:
      tags: ["Groups"]
      summary: Set group name
      operationId: setGroupName
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  description: New group name
                  minLength: 3
                  maxLength: 30
                  pattern: "^.*$"
      responses:
        '200':
          description: Name updated
          content:
            application/json:
              schema:
                type: object
                description: Success
        '500':
          $ref: '#/components/responses/InternalServerError'

  /groups/{groupId}/photo:
    parameters:
      - $ref: '#/components/parameters/GroupId'
    put:
      tags: ["Groups"]
      summary: Set group photo
      operationId: setGroupPhoto
      requestBody:
        required: true
        content:
          image/png:
            schema:
              type: string
              format: binary
              description: Photo data
              minLength: 1
              maxLength: 10485760
              pattern: "^.*$"
      responses:
        '200':
          description: Photo updated
          content:
            application/json:
              schema:
                type: object
                description: Success
        '500':
          $ref: '#/components/responses/InternalServerError'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Enter your userId here (from /session)

  parameters:
    ConversationId:
      name: conversationId
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 64
        pattern: "^[a-zA-Z0-9_-]+$"
    MessageId:
      name: messageId
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 64
        pattern: "^[a-zA-Z0-9_-]+$"
    GroupId:
      name: groupId
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 64
        pattern: "^[a-zA-Z0-9_-]+$"

  schemas:
    ErrorResponse:
      type: object
      description: Standard error wrapper
      properties:
        message:
          type: string
          description: Error details
          minLength: 1
          maxLength: 500
          pattern: "^.*$"

    ConversationSummary:
      type: object
      description: Summary for list view
      properties:
        conversationId:
          type: string
          description: ID of conversation
          minLength: 1
          maxLength: 64
          pattern: "^[a-zA-Z0-9_-]+$"
        name:
          type: string
          description: Name of user or group
          minLength: 1
          maxLength: 50
          pattern: "^.*$"
        photo:
          type: string
          format: byte
          description: Base64 preview or null
          minLength: 0
          maxLength: 100000
          pattern: "^.*$"
        lastMessage:
          $ref: '#/components/schemas/Message'
        isGroup:
          type: boolean
          description: True if group chat
        unreadCount:
          type: integer
          description: Number of unread messages

    ConversationDetail:
      type: object
      description: Full conversation details
      properties:
        conversationId:
          type: string
          description: ID
          minLength: 1
          maxLength: 64
          pattern: "^[a-zA-Z0-9_-]+$"
        name:
          type: string
          description: Name
          minLength: 1
          maxLength: 50
          pattern: "^.*$"
        members:
          type: array
          description: List of member IDs
          items:
            type: string
            minLength: 1
            maxLength: 64
            pattern: "^[a-zA-Z0-9_-]+$"
        messages:
          type: array
          description: List of messages
          items:
            $ref: '#/components/schemas/Message'

    Message:
      type: object
      description: A single message
      properties:
        messageId:
          type: string
          description: Unique ID
          minLength: 1
          maxLength: 64
          pattern: "^[a-zA-Z0-9_-]+$"
        senderId:
          type: string
          description: Who sent it
          minLength: 1
          maxLength: 64
          pattern: "^[a-zA-Z0-9_-]+$"
        content:
          type: string
          description: Text content
          minLength: 0
          maxLength: 2000
          pattern: "^.*$"
        timestamp:
          type: string
          format: date-time
          description: When it was sent
          minLength: 1
          maxLength: 50
          pattern: "^.*$"
        status:
          type: integer
          description: 1=received, 2=read
          enum: [1, 2]
        reactions:
          type: array
          description: Reactions on this message
          items:
            type: object
            properties:
              user:
                type: string
                description: Who reacted
                minLength: 1
                maxLength: 64
                pattern: "^[a-zA-Z0-9_-]+$"
              emoji:
                type: string
                description: The emoji
                minLength: 1
                maxLength: 4
                pattern: "^.*$"

  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'